Remove RTCIceServer API from BCD
This PR removes the `RTCIceServer` API from BCD. This feature is a type (ex. a dictionary, enum, mixin, constant or WebIDL typedef) that we have explicitly stated not to document separately from the feature(s) that depend on it, as they are virtually invisible to the end developer.

Tests Used: https://mdn-bcd-collector.gooborg.com/tests/api/RTCIceServer

Additional Notes: I'd love for someone to take on the content changes if possible, but if nobody is interested I'd be happy to give the content updates a shot!