Update Firefox data for MerchantValidationEvent API
This PR updates and corrects version values for Firefox and Firefox Android for the `MerchantValidationEvent` API. This replaces the version number with "preview" as it is only available in nightly builds of Firefox and should not have a version number associated accordingly.  Flag data is also propogated down to subfeatures as per our guidelines.