Update Firefox data for SpeechSynthesisVoice API
This PR updates and corrects version values for Firefox and Firefox Android for the `SpeechSynthesisVoice` API. This bug was discovered when writing tests for the mdn-bcd-collector. This adds a note mentioning this bug and recommending a solution based on https://github.com/GooborgStudios/mdn-bcd-collector/commit/365ad25c361516fc26340f3946884ff90c85f342.