Add missing features for javascript.builtins.Intl.NumberFormat (#20588)
This PR is a part of a project to add missing interfaces and interface features to BCD that are from an active spec (including WICG specs) and is supported in at least one browser. This particular PR adds the missing features of the `NumberFormat` member of the `Intl` JavaScript builtin. This fixes #6046 by adding the missing arguments.  Note: all data was simply mirrored from the main constructor's data, assuming they are all features that were supported since the constructor was introduced.