Remove irrelevant Firefox flag data for api.HTMLMediaElement.preservesPitch
This PR removes irrelevant flag data for Firefox and Firefox Android for the `preservesPitch` member of the `HTMLMediaElement` API as per the corresponding [data guidelines](https://github.com/mdn/browser-compat-data/blob/main/docs/data-guidelines/index.md#removal-of-irrelevant-flag-data). This PR was created from results of the `remove-redundant-flags` script.

Additional Notes: This fixes #20766 in the meantime.