Add missing api.OffscreenCanvas.getContext.webgpu_context feature
This PR is a part of a project to add missing interfaces and interface features to BCD that are from an active spec (including WICG specs) and is supported in at least one browser. This particular PR adds the missing `getContext.webgpu_context` member of the `OffscreenCanvas` API. The data was mirrored from api.HTMLCanvasElement.getContext.webgpu_context.  This fixes #9815.