Update Chromium data for javascript.builtins.Intl.DateTimeFormat.DateTimeFormat.options_parameter.options_timeZoneName_parameter.extended_values
This PR updates and corrects version values for Chromium (Chrome, Opera, Samsung Internet, WebView Android) for the `DateTimeFormat.DateTimeFormat.options_parameter.options_timeZoneName_parameter.extended_values` member of the `Intl` JavaScript builtin. This fixes #14586, which contains the supporting evidence for this change.

Additional Notes: This also renames the feature to better represent what it is (it's not IANA time zone names).