Update Safari iOS data for caret-color CSS property
This PR updates and corrects version values for Safari iOS/iPadOS for the `caret-color` CSS property. This fixes #15128, which contains the supporting evidence for this change.

Additional Notes: The behavior described had not been seen in any version of Safari iOS I had tested.  Additionally, the commit that fixed the bug landed in iOS 11.3.