Update Chromium data for api.Document.domain
This PR updates and corrects version values for Chromium (Chrome, Opera, Samsung Internet, WebView Android) for the `domain` member of the `Document` API. This updates the note introduced in #18500 to a) better explain the new behavior, b) fix the version number (see https://chromestatus.com/feature/5428079583297536), c) remove an empty link to a demo, and d) link to an MDN article explaining origin-keyed agent clusters, not a ChromeStatus page.

Additional Notes: This fixes #21221.