Update Safari data for css.properties.user-select.text
This PR updates and corrects version values for Safari (Desktop and iOS/iPadOS) for the `text` member of the `user-select` CSS property. This fixes #16566, which contains the supporting evidence for this change.

Additional Notes: Although the note doesn't make too much sense, I can't type or add any content in Safari 11.1+.