Update Safari data for html.global_attributes.hidden
This PR updates and corrects version values for Safari (Desktop and iOS/iPadOS) for the `hidden` member of the `global_attributes` HTML feature. This fixes #16619, which contains the supporting evidence for this change.

Additional Notes: The reporter states this is an iOS issue, but it affects Safari Desktop as well.