Add stroked HTML text compatibility info for CSS `paint-order` property (#21426)
Browser support for CSS `paint-order` on HTML text differs between
browsers. This is because as originally specified, the property had a
restriction saying that it should only apply to SVG. That restriction
has since been removed, see https://github.com/w3c/fxtf-drafts/issues/107

Support for `paint-order` on HTML text was included along with the
initial support for `paint-order` in Firefox 60, see
https://bugs.webkit.org/show_bug.cgi?id=168601

Support for `paint-order` on HTML text was added to WebKit in
https://github.com/WebKit/WebKit/commit/73a116fe7ce0dffc4856bb8dc3986e49308df943
which was included in WebKit 604. According to browsers/safari.json,
this corresponds to Safari 11.

Support for `paint-order` on HTML text in Chromium is still pending:
https://bugs.chromium.org/p/chromium/issues/detail?id=815111

Fixes #17951
Fixes #19154