Update Safari data for html.global_attributes.hidden.until-found_value
This PR updates and corrects version values for Safari (Desktop and iOS/iPadOS) for the `hidden.until-found_value` member of the `global_attributes` HTML feature. This fixes #The Firefox bug was not added as was mentioned in the issue, because it has been unconfirmed since it was opened two years ago., which contains the supporting evidence for this change.