Add missing javascript.builtins.Error.cause.displayed_in_console feature (#21570)
This PR is a part of a project to add missing interfaces and interface features to BCD that are from an active spec (including WICG specs) and is supported in at least one browser. This particular PR adds the missing `cause.displayed_in_console` member of the `Error` JavaScript builtin. This fixes #19004, which contains the supporting evidence for this change.