Add missing css.selectors.selection.text-decoration feature (#21536)
This PR is a part of a project to add missing interfaces and interface features to BCD that are from an active spec (including WICG specs) and is supported in at least one browser. This particular PR adds the missing `text-decoration` member of the `selection` CSS selector. The data comes from manual testing, running test code through BrowserStack, SauceLabs and custom VMs.

Test Code:

```html
<style>
    p::selection {
        color: seagreen;
        background-color: white;
        text-decoration: underline;
    }
</style>

<p>Lorem ipsum dolor sit amet</p>
```

Fixes #17817.