Add missing css.types.length.em feature (#21488)
* Add missing css.types.length.em feature

This PR is a part of a project to add missing interfaces and interface features to BCD that are from an active spec (including WICG specs) and is supported in at least one browser. This particular PR adds the missing `em` member of the `length` CSS value type. This mirrors the data from `rem` (which is probably accurate enough for our needs).  Fixes #15293.

* Update length.json

* Remove IE (instead of leaving inaccurate data)

* Fix typo

* IE must be present in data

* Use IE 3

---------

Co-authored-by: Florian Scholz <fs@florianscholz.com>