Update all browsers data for css.types.image.*-gradient (#21472)
This PR updates and corrects version values for all browsers for the `*-gradient` member of the `image` CSS value type. This adds a `premultiplied_gradients` subfeature to all three gradient types, as all browsers (not just Firefox) had this issue in the past.  In a way, this fixes #14598.

Test Code:
```html
<div id="test"></div>
<style>
	body {
		background-color: black;
	}
	#test {
		height: 100vh;
		background: linear-gradient(to bottom, red, rgba(255, 255, 255, 0));
	}
</style>
```

If the gradient started to transition into white, then premultiplication was not performed.  If the gradient faded perfectly to black, premultiplication was performed properly.