Remove MediaTrack dictionaries from BCD
This PR removes the `MediaTrackConstraints`, `MediaTrackSettings`, and `MediaTrackSupportedConstraints` dictionaries from BCD. These are a type (ex. a dictionary, enum, mixin, constant or WebIDL typedef) that we have explicitly stated not to document separately from the feature(s) that depend on it, as they are virtually invisible to the end developer.