Remove RTCRtpSynchronizationSource API from BCD (#21759)
This PR removes the `RTCRtpSynchronizationSource` API from BCD. This feature is a type (ex. a dictionary, enum, mixin, constant or WebIDL typedef) that we have explicitly stated not to document separately from the feature(s) that depend on it, as they are virtually invisible to the end developer.

Additional Notes: The MDN pages have already been removed.