html.img: Remove phantom onerror deprecation
The `onerror` attribute to the `img` element has been marked
deprecated since the table data was introduced in 061483a.
However, that deprecation was never sourced to any specs or
references, and multiple inquiries since have failed to uncover
any explanation for it. It appears to have been a simple error.