Mark the item* microdata attributes as not supported by browsers
The API reflecting these attributes was removed already:
https://github.com/mdn/browser-compat-data/pull/16226

The attributes have never had any effect beyond the API for accessing
them, so it's fair to say these attributes are not supported.

A search for "itemscope" in the source of Chromium, Gecko and WebKit
finds only old tests and similar, no code handling the attribute.