Bump web-specs from v2.41.0 to v3.0.0 and adjust test accordingly (#22185)
A new major of web-specs was released. Breaking change is that the presence of
a `nightly` property is no longer guaranteed. This happens for specs that do
not have a public URL, such as ISO specs.

The test that checks spec URLs needed to be adjusted accordingly.

Also, note the new version includes the Threading proposal for the WASM JS API,
which could thus be dropped from the list of exceptions.