Update Safari data for Content-Security-Policy HTTP header (#22521)
This PR updates and corrects version values for Safari (Desktop and iOS/iPadOS) for the `Content-Security-Policy` HTTP header. The data comes from a commit in the browser's source code, mapped to a version number using available tooling or via the commit timestamp.

Commit: https://github.com/WebKit/WebKit/commit/e6ad7c23647cc760606c1885492ba4e77ed7d986

Additional Notes: Fixes #22181.  Note: `script-src-elem` was set to 16.1 in #18911, but looking at the release notes, 16.1 fixed the CSP in workers, not that it introduced it.