Split `api.Headers.lexicographical_sorting` in two and update the data
I was investingating this feature in the course of creating a fetch feature in web-features. The data for Safari looked dubious. I soon found out that the two parts of this (sorting and combining) have different histories. Instead of using partial implementations, I split the feature in two.

The highlights here:

- Safari always supported both
- Edge acquired them separately
- Firefox got both at the same time, but later than reported