Remove all "Thing() without new throws" entries
These entries document the absence of a bug. If inverted, they describe
a feature which was supported, but has long been removed. Such entries
are irrelevant per this guidline:
https://github.com/mdn/browser-compat-data/blob/main/docs/data-guidelines/index.md#removal-of-irrelevant-features

Another way to look at this is that the standard behavior works before
and after this bug was fixed, and the non-standard beahvior is
additional API surface which no longer exists.