Correct and simplify Firefox data for image-set()
https://bugzil.la/1832901 is about `content: linear-gradient(...)`, but
`content: image-set(...)` works in Firefox 88 in this test:
https://software.hixie.ch/utilities/js/live-dom-viewer/?saved=12629

That `type()` didn't work in Firefox 88 was confirmed with this test:
https://software.hixie.ch/utilities/js/live-dom-viewer/?saved=12631

Use partial_implementation for that single-version issue.

Support for `content: -webkit-image-set()` starting in Firefox 90 was
confirmed with additional manual testing (tweaking above tests).

https://bugzil.la/1696314 is only about the `cursor` property, the
`content` property was fixed in Firefox 113:
https://bugzilla.mozilla.org/show_bug.cgi?id=1684958

The exact scope of the changes in Firefox 113 is hard to discern, but
since `content: image-set(...)` works and `cursor: image-set(...)` isn't
in the spec, there's nothing worth noting.