Remove HTTP/1.1 tags
Computing the web-features status from these did not work out:
https://github.com/web-platform-dx/web-features/pull/946

As a category for what was in HTTP/1.1 the tags are probably mostly
correct, but because of all of the true values we can't compute anything
from BCD, and also not show any Baseline badge on MDN with confidence.

Hopefully these tags will make a comeback in the future.