Update public_class_fields for Safari
Marks Safari 14-15 as having partial support for public class fields. While they do have a basic implementation, there is a critical bug which leads to exceptions in many cases. The exceptions are especially hard to identify/debug because they are resolved when dev-tools are open.

The bug is resolved in Safari 16.

References:

- Safari bug: https://bugs.webkit.org/show_bug.cgi?id=236843

- `babel/babel` discussion regarding adding 'bugfix' behavior for this issue: https://github.com/babel/babel/issues/14289