Update checkVisibility() parameter support
The parameters for this method were changed in the spec, and all
implementations proceeded in the same way:

 - First checkVisibility() with checkOpacity and checkVisibilityCSS
 - Then contentVisibilityAuto, opacityProperty, and visibilityProperty,
   making the original parameters historical aliases.

Chromium:

 105: https://chromiumdash.appspot.com/commit/60a7831496032f7b132bc96ec62542150ecb1103
 121: https://chromiumdash.appspot.com/commit/21e42c2ba69d45e08009432c56a5327046c499e6

Gecko:

 106: https://bugzilla.mozilla.org/show_bug.cgi?id=1777293#c4
 122: https://bugzilla.mozilla.org/show_bug.cgi?id=1859852#c6

WebKit:

 618.1.3: https://github.com/WebKit/WebKit/commit/7d04a9c1b62bc8a8101c390f6a164cf4b3da9cfa
 618.1.5: https://github.com/WebKit/WebKit/commit/2d9a96581fafa907ffd9ddfb4d8ca2bb24ae0692

Based on the WebKit version, both ended up in Safari 17.4. Support for
all 5 parameters in Safari 17.5 was confirmed with a dev instance of the
collector with the tests fixed:
https://github.com/openwebdocs/mdn-bcd-collector/pull/1478

Remove the web-features tags from the parameters as it is not clear how
thiss will be handled in web-features yet.