Update Chrome/Safari data for html.elements.base.href.relative_url
This PR updates and corrects version values for Chrome and Safari for the `href.relative_url` member of the `base` HTML element. This sets the version to be a range corresponding to the date the feature was first tracked in BCD (1e76c2ec4001b790ecfee95815c8b72663e47ee9).