Remove html.elements.input.x-moz-errormessage from BCD
This PR removes the `x-moz-errormessage` member of the `input` HTML element from BCD. Per the [data guidelines](https://github.com/mdn/browser-compat-data/blob/main/docs/data-guidelines/index.md#removal-of-irrelevant-features), this feature can be considered irrelevant and may be removed from BCD accordingly. Even if the current data suggests that the feature is supported, lack of support has been confirmed by the [mdn-bcd-collector](https://mdn-bcd-collector.gooborg.com) project (v10.10.7).

Tests Used: https://mdn-bcd-collector.gooborg.com/tests/html/elements/input/x-moz-errormessage

Additional Notes: There is a message about this attribute in the content, but it simply states that it used to exist and was removed, so it doesn't need to be deleted.