Update Chrome/Safari data for mixed-content HTTP feature
This PR updates and corrects version values for Chrome and Safari for the `mixed-content` HTTP feature. This PR was intended to be a part of the true->ranged project with ranges based upon the commit the feature was added in, but the feature was first added to BCD this year (#23220), so this gives the roughest guess based upon all the sibling data.