Remove `api.MouseEvent.region`
This seems to be a non-feature. I could find no reference to it in a
current spec, no open Chromium bugs relating to it, and failed to find
evidence of it under inspection in Chrome with the flag turned on.

The docs were removed from MDN in 2022:
https://github.com/mdn/content/pull/17743.

The data for this property probably ought to have been removed alongside
https://github.com/mdn/browser-compat-data/pull/8442.