Update Chrome Android data for GPU API
This PR updates and corrects version values for Chrome Android for the `GPU` API. This fixes #22159, which contains the supporting evidence for this change.

Additional Notes: While the collector currently doesn't have results for Android browsers, the data was confired in the collector by manually running the tests on Chrome Android via BrowserStack.