Update Firefox data for javascript.statements.import.service_worker_support (#23555)
* Update Firefox data for javascript.statements.import.service_worker_support

This PR updates and corrects version values for Firefox and Firefox Android for the `import.service_worker_support` member of the `statements` JavaScript feature. This fixes #23554, which contains the supporting evidence for this change.

Additional Notes: It is worth noting that this feature, and a subfeature of `ServiceWorker` indicating ECMAScript module support, were both added in #19835. It has been confirmed that ECMAScript modules aren't supported in Service Workers in Firefox yet.

* Add impl_url

---------

Co-authored-by: Florian Scholz <fs@florianscholz.com>