`css.types.image.gradient`: Remove `-ms-` prefix
An abstract type having a prefix doesn't make any sense. Nevertheless, I
tested gradients on Edge 18 and 15 and found that prefixes had no effect
for `conic-gradient()` or `linear-conic-gradient()` and neither was
supported in any case.

I believe this be a case of a prefix being wrongly reported, then
copied, from IE data.