Add missing css.properties.object-position.applies_to_iframe_elements feature
This PR adds the missing `applies_to_iframe_elements` member of the `object-position` CSS property. This fixes #23586, which contains the supporting evidence for this change.

Additional Notes: The test code was copied and adapted to a local server, then manually run on Firefox 36 and old Chrome/Safari versions to confirm support.