Clarify notes about clipboard permissions
The notes about version 85 appear to be about this change:
https://chromestatus.com/feature/5767075295395840

However, there is no requirement to use the feature policy or permission
policy, that is only used to constrain a page more than the default,
typically to disallow requesting a permission at all.

It was confirmed that the two permissions go back to Chrome 66, by
testing `navigator.permissions.query({name: "clipboard-read"})` and
"clipboard-write" in DevTools in Chrome 66.

The only remaining noteworthy thing is the behavior around user
activation. The existing claims in the notes weren't double checked,
just carried forward while updating permissions language.

Drive-by: fix surpressed typo in Firefox notes