Update slightly complex Chromium data for webextensions.api
This PR updates and corrects version values for Chromium (Chrome, Opera, Samsung Internet, WebView Android) for the `api` member of the `webextensions` Web Extensions interface. This sets the feature(s) to a version range based upon the date that the feature was added to BCD with the intent of replacing `true` values with ranged values to eliminate `true` values from BCD.

Commit/PR Adding the Feature: #166

This is a cherry-pick of the files containing more complicated changes that were originally in #23450.  While most of the features are set to ≤58, some were set to different version numbers to satisfy the consistency linter, or there was other readily available information already in BCD to confirm the earlier version number.