Update Safari data for font-variant-emoji CSS property (#23659)
This PR updates and corrects version values for Safari (Desktop and iOS/iPadOS) for the `font-variant-emoji` CSS property. This fixes #23216, which contains the supporting evidence for this change.

Additional Notes: The version number may not be entirely accurate, but since flag data really only matters for the current version, I figure that it's not too important to track down.