Update initial support for <iframe sandbox> in WebKit
This was implemented without a flag, and with the reflected IDL
attribute, in WebKit 532.6:
https://github.com/WebKit/WebKit/commit/c205c4e7d40da38c5009df548c2316a358a4fca7
https://github.com/WebKit/WebKit/blob/c205c4e7d40da38c5009df548c2316a358a4fca7/WebCore/Configurations/Version.xcconfig

That maps to Chrome 5 + Safari 5 + iOS 4, which also matches exactly
earlier updates based on collector test results:
https://github.com/mdn/browser-compat-data/pull/7243
https://github.com/mdn/browser-compat-data/pull/7290
https://github.com/mdn/browser-compat-data/pull/7402

The handled attribute tokens were "allow-same-origin", "allow-forms",
and "allow-scripts", so update that data to match.

The "allow-popups" entry was bumped to pass the lint.