Update Chrome/Safari versions for sandbox="allow-top-navigation"
This was implemented in WebKit 533.3:
https://github.com/WebKit/WebKit/commit/f4228dc0076684651c8b54680db4d1d7b964a321
https://github.com/WebKit/WebKit/blob/f4228dc0076684651c8b54680db4d1d7b964a321/WebCore/Configurations/Version.xcconfig

That maps to Chrome 5 or 6, Safari 5, and iOS 4.2.

Chrome 5 or 6 is ambiguous because the engine version is 533 with no
minor version. Chrome 6 is 534.3, so it was definitely in Chrome 6.

Unlike other sandbox data, mirroring can be used for iOS, as Safari 5
does map to iOS 4.2.