Update Firefox data for events Web Extensions interface
This PR updates and corrects version values for Firefox and Firefox Android for the `events` Web Extensions interface. This sets the feature(s) to a version range based upon the date that the feature was added to BCD with the intent of replacing `true` values with ranged values to eliminate `true` values from BCD.

Commit/PR Adding the Feature: #166, #1686

Additional Notes: The "≤50" comes from a subfeature that's set as "50".