Add missing web-features fields for webview_ios
I tried to generate web-features locally using WebViews
as part of the browser set.

I found that there were a few keys that webview_ios
wasn't set for.

As far as I can see, in each of the places it was missing
it was set as mirror for other keys for those elements.