WebView iOS also supports css.types.color.color-mix (#26093)
Update WebView iOS data for css.types.color.color-mix

This PR updates and corrects version values for WebView iOS/iPadOS for the `color-mix` member of the `color` CSS value type. This fixes #25451, which contains the supporting evidence for this change.

Additional Notes: The collector has a typo which is causing the test to not pass.  This will be fixed.