Remove api.MLOperand.MLNumber from BCD (#26138)
This PR removes the `MLNumber` member of the `MLOperand` API from BCD. This feature is a type (ex. a dictionary, enum, mixin, constant or WebIDL typedef) that we have explicitly stated not to document separately from the feature(s) that depend on it, as they are virtually invisible to the end developer.