Mark the Iterator() constructor as not supported
The data comes from these PRs:
https://github.com/mdn/browser-compat-data/pull/21998
https://github.com/mdn/browser-compat-data/pull/24332
https://github.com/mdn/browser-compat-data/pull/26285

In Chrome and Node.js, invoking `new Iterator()` throws a TypeError with the message "Abstract class Iterator not directly constructable". In Firefox, the message is "Iterator constructor can't be used directly".

Since the constructor can't be used directly, saying that it is supported seems unhelpful. Possibly the entry should be removed.