ci(workflows): add system-file-changes workflow (#26853)
Adds new `system-file-changes` workflow that fails when anyone who isn't an admin 
or a BCD owner changes any file that isn't BCD JSON data or Markdown docs.

Prevents accidental or malicious changes, and ensures that these get noticed.

This is more relevant now with the PR Review Companion, which could lead to
reviewers no longer checking all changes on the "Files changed" tabs, especially
if the list of changes or changed files is long, and doesn't fit on the screen.