feat(browsers): add Bun to BCD (#23440)
This PR adds the Bun runtime to BCD, and tracks the data.

The initial data was generated with the help of the runtime-compat
project, a project that uses the mdn-bcd-collector to collect test
results for JavaScript runtimes.

The release data was obtained by a script written by @alii, which
has been integrated into the release updater workflow.

Special thanks to the Bun team for all of their assistance!


Co-authored-by: Florian Scholz <fs@florianscholz.com>
Co-authored-by: Alistair Smith <hi@alistair.sh>
Co-authored-by: Claas Augner <caugner@mozilla.com>