Update Safari data for ExtendableCookieChangeEvent API
This PR updates and corrects version values for Safari (Desktop and iOS/iPadOS) for the `ExtendableCookieChangeEvent` API. The mdn-bcd-collector reported this to be unsupported in Safari 18.4 through 18.6, even though it was reported as supported in Safari 18.4 beta. After doing some additional digging, it appears that this feature was hidden behind a flag that must have been enabled in 18.4 beta but disabled in stable releases.