Add many missing CSS property values (early browser versions) (#28265)
This PR is a part of a process to add all of the missing spec-defined CSS property values. These values are added by the mdn-bcd-collector, following an upgrade of Webref CSS to v7 and directly parsing the syntax for each property.

This specific PR focuses on CSS property values that were added in early browser versions. If a property had a prefix or alternative name, and the collector reported that a value was supported in the first version that supported the standard name, then it was assumed that the value was supported from the beginning and had been updated accordingly.

Spec URLs were obtained for as many values as possible.  However, once again, csswg.org went down frequently during the creation of this PR. Some spec URLs could not be obtained while the website was down. (Last property checked: `filter`)

Notes:
- Chrome: Chrome 15+ was tested, with only a few versions that failed to be tested.
- Firefox: Firefox 5 through 49 crashed during tests, so no results were obtained for those versions. Versions within this range were guesstimated by parent data, with the intent to fine-tune the version numbers later if needed.
- Safari: Not all Safari versions were tested during this process (partially due to a lack of VM access), but test results were collected for every major Safari version. Specific minor versions were inferred by parent data, with the intent to fine-tune the version numbers later if needed.

---------

Co-authored-by: Claas Augner <caugner@mozilla.com>